Integrate ShiftCalendar into CreatePostingShifts
